ENH: reusing code :D
